[#150] Refactor: 가입기간 -> 일기작성기간 수정
