+++ src/main/java/umc/GrowIT/Server/repository/ChallengeRepository/ChallengeRepository.java
@@ -1,4 +1,4 @@
-package umc.GrowIT.Server.repository;
+package umc.GrowIT.Server.repository.ChallengeRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
