+++ src/main/java/umc/GrowIT/Server/service/ChallengeService/ChallengeCommandServiceImpl.java
@@ -7,11 +7,8 @@ import umc.GrowIT.Server.apiPayload.code.status.ErrorStatus;
 import umc.GrowIT.Server.apiPayload.exception.ChallengeHandler;
 import umc.GrowIT.Server.apiPayload.exception.UserHandler;
 import umc.GrowIT.Server.converter.ChallengeConverter;
-import umc.GrowIT.Server.domain.Challenge;
-import umc.GrowIT.Server.domain.User;
 import umc.GrowIT.Server.domain.UserChallenge;
-import umc.GrowIT.Server.repository.ChallengeRepository;
-import umc.GrowIT.Server.repository.UserChallengeRepository;
+import umc.GrowIT.Server.repository.ChallengeRepository.UserChallengeRepository;
 import umc.GrowIT.Server.repository.UserRepository;
 import umc.GrowIT.Server.service.ImageService.ImageService;
 import umc.GrowIT.Server.web.dto.ChallengeDTO.ChallengeRequestDTO;
