+++ src/main/java/umc/GrowIT/Server/service/ChallengeService/ChallengeQueryServiceImpl.java
@@ -6,22 +6,20 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.GrowIT.Server.apiPayload.code.status.ErrorStatus;
 import umc.GrowIT.Server.apiPayload.exception.ChallengeHandler;
 import umc.GrowIT.Server.converter.ChallengeConverter;
-import umc.GrowIT.Server.domain.Challenge;
-import umc.GrowIT.Server.domain.ChallengeKeyword;
 import umc.GrowIT.Server.domain.Keyword;
 import umc.GrowIT.Server.domain.UserChallenge;
 import umc.GrowIT.Server.domain.enums.UserChallengeType;
 import umc.GrowIT.Server.repository.ChallengeKeywordRepository;
-import umc.GrowIT.Server.repository.ChallengeRepository;
+import umc.GrowIT.Server.repository.ChallengeRepository.ChallengeRepository;
 import umc.GrowIT.Server.repository.KeywordRepository;
-import umc.GrowIT.Server.repository.UserChallengeRepository;
+import umc.GrowIT.Server.repository.ChallengeRepository.UserChallengeRepository;
+import umc.GrowIT.Server.repository.diaryRepository.DiaryRepository;
 import umc.GrowIT.Server.web.dto.ChallengeDTO.ChallengeResponseDTO;
 
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -31,6 +29,7 @@ public class ChallengeQueryServiceImpl implements ChallengeQueryService {
     private final UserChallengeRepository userChallengeRepository;
     private final KeywordRepository keywordRepository;
     private final ChallengeKeywordRepository challengeKeywordRepository;
+    private final DiaryRepository diaryRepository;
     @Override
     public int getTotalCredits(Long userId) {
         // TODO: 총 크레딧 수 조회 로직 구현
@@ -46,15 +45,16 @@ public class ChallengeQueryServiceImpl implements ChallengeQueryService {
 
     @Override
     public String getUserDate(Long userId) {
-        // TODO: 사용자가 가입한 날짜와 오늘 날짜 간의 차이를 계산
-        LocalDate joinDate = challengeRepository.findJoinDateByUserId(userId)
-                .orElse(LocalDate.now()); // 가입 날짜가 없으면 오늘 날짜를 기본값으로 사용
+        // TODO: 사용자가 일기를 처음 작성한 날짜와 오늘 날짜 간의 차이를 계산
+        LocalDate firstDiaryDate = diaryRepository.findFirstDiaryDateByUserId(userId) // 최초 일기 작성 날짜 조회
+                .orElse(LocalDate.now()); // 일기를 작성한 기록이 없으면 오늘 날짜를 기본값으로 사용
 
         // 오늘 날짜와의 차이를 계산
-        long days = ChronoUnit.DAYS.between(joinDate, LocalDate.now());
+        long days = ChronoUnit.DAYS.between(firstDiaryDate, LocalDate.now());
 
         return "D+" + days;
     }
+
 
     @Override
     public ChallengeResponseDTO.ChallengeHomeDTO getChallengeHome(Long userId) {
