+++ src/main/java/umc/GrowIT/Server/web/dto/ChallengeDTO/ChallengeRequestDTO.java
@@ -15,7 +15,7 @@ public class ChallengeRequestDTO {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class ProofRequestDTO {
-        @Schema(description = "인증 이미지 파일 (multipart/form-data)", nullable = true, requiredMode = Schema.RequiredMode.REQUIRED)
+        @Schema(description = "인증 이미지 파일 (multipart/form-data)", requiredMode = Schema.RequiredMode.REQUIRED)
         private MultipartFile certificationImage;
         @Schema(description = "소감 (텍스트)", requiredMode = Schema.RequiredMode.REQUIRED)
         private String thoughts;
